Move library test to appropriate file
